StorageAI/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
description = "Hands-free inventory assistant with PyQt5 UI"
version = "0.1.0"
name = "storageai"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PyQt5>=5.15.11",
  "openai>=1.40.0",
  "pinecone-client>=3,<4",
  "sounddevice>=0.4.6",
  "soundfile>=0.12.1",
  "numpy>=1.26",
  "pydantic>=2.7",
  "requests>=2.32.3",
  # Notion removed; using Airtable REST via requests
  "python-dateutil>=2.9.0.post0",
  "sqlite-utils>=3.36",
]
